// frontend/src/app/models/projeto.ts
export interface Projeto {
   id?:                any;
   dataAbertura?:   string;
   dataFechamento?: string;
   prioridade:      string;
   status:          string;
   titulo:          string;
   observacoes:     string;
   tecnico:            any;
   cliente:            any;
   nomeCliente:     string;
   nomeTecnico:     string;
}